+++ src/Git_Updater/REST/Rest_Update.php
@@ -124,8 +124,14 @@ class Rest_Update {
 			1
 		);
 
-		// Add authentication header to download package.
-		add_filter( 'http_request_args', [ Singleton::get_instance( 'Fragen\Git_Updater\API\API', $this ), 'download_package' ], 15, 2 );
+		// Load hook for adding authentication headers for download packages.
+		add_filter(
+			'upgrader_pre_download',
+			function () {
+				add_filter( 'http_request_args', [ Singleton::get_instance( 'Fragen\Git_Updater\API\API', $this ), 'download_package' ], 15, 2 );
+				return false; // upgrader_pre_download filter default return value.
+			}
+		);
 
 		$upgrader = new \Plugin_Upgrader( $this->upgrader_skin );
 		$upgrader->upgrade( $plugin->file );
@@ -190,8 +196,14 @@ class Rest_Update {
 			1
 		);
 
-		// Add authentication header to download package.
-		add_filter( 'http_request_args', [ Singleton::get_instance( 'Fragen\Git_Updater\API\API', $this ), 'download_package' ], 15, 2 );
+		// Load hook for adding authentication headers for download packages.
+		add_filter(
+			'upgrader_pre_download',
+			function () {
+				add_filter( 'http_request_args', [ Singleton::get_instance( 'Fragen\Git_Updater\API\API', $this ), 'download_package' ], 15, 2 );
+				return false; // upgrader_pre_download filter default return value.
+			}
+		);
 
 		$upgrader = new \Theme_Upgrader( $this->upgrader_skin );
 		$upgrader->upgrade( $theme->slug );
@@ -283,7 +295,7 @@ class Rest_Update {
 				'elapsed_time' => round( ( microtime( true ) - $start ) * 1000, 2 ) . ' ms',
 				'deprecated'   => $deprecated,
 			];
-			$this->log_exit( $http_response, 417 );
+			$this->log_exit( $http_response, 418 );
 		}
 
 		// Only set branch on successful update.
@@ -313,7 +325,7 @@ class Rest_Update {
 
 		if ( $this->is_error() ) {
 			$response['success'] = false;
-			$this->log_exit( $response, 417 );
+			$this->log_exit( $response, 418 );
 		}
 		$this->log_exit( $response, 200 );
 	}