+++ CHANGES.md
@@ -1,5 +1,14 @@
 #### [unreleased]
 
+#### 12.7.0 / 2024-11-30
+* fix missing/incorrect textdomains
+* look for `__()` functions loading in hooks before `init`
+* remove `load_plugin_textdomain()`
+* add git host icon to single site theme description
+* don't save to GitHub.com access token from single repo remote install
+* fix PHP 8.1 creation dynamic variable from `class REST_API`
+* update REST API response to return `plugins_api()` or `themes_api()` style response
+
 #### 12.6.0 / 2024-10-13
 * check existence of `FS__RESOLVE_CLONE_AS` before setting
 * add filter hook `gu_api_domain` to set domain for default API updating